+++ src/Game.php
@@ -9,6 +9,15 @@ class Game
 {
     public $id;
     public $title;
+    public $forumTopicId;
     public $consoleId;
     public $imageIcon;
+    public $gameIcon;
+    public $imageTitle;
+    public $imageInGame;
+    public $imageBoxArt;
+    public $publisher;
+    public $developer;
+    public $genre;
+    public $releaseDate;
 }